Treat "multi-lets" as syntactic sugar
Removes any mention of "multi-lets" from the semantics, while adding a
note to the grammar that `let... let... in...` is to be treated as
syntactic sugar for `let... in let... in...`. The resulting standard is
equivalent to the old one, only arguably simpler and easier to follow.

Addresses dhall-lang#615; see
dhall-lang@022105e
for the commit that originally introduced multi-lets into the standard.

This time with fewer unrelated whitespace changes.